assertion for tuple; complete get_conn_tets
