+++ src/EdgeCollapsingConn.cpp
@@ -1,0 +1,13 @@
+//
+// Created by Yixin Hu on 12/9/21.
+//
+#include <wmtk/TetMesh.h>
+
+#include <wmtk/TupleUtils.hpp>
+
+bool wmtk::TetMesh::collapse_edge(const Tuple& t, std::vector<Tuple>& new_edges)
+{
+    // todo
+
+    return true;
+}
